+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.4.10
+
+- **FEAT**: Dalgeurak Service - applicationTeacherMealCancel Function
+- **FIX** - Dalgeurak Service - Meal Price API URL
+
 ## 0.4.9
 
 - **FEAT**: Dalgeurak Service - setStayMealPrice, getStayMealPrice Function